# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API which provide you backend service for your minimalist ecommerce app",
        "title": "Anya Day API",
        "contact": {
            "name": "Developer",
            "email": "riidloa@gmail.com"
        },
        "version": "beta"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/dev/users/{id}/address": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update selected address field of an existing user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] update address of an existing user.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update address field you like. Remove that you won't.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddressInputNotBinding"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.NormalResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Create fresh complete address of an existing user. If you want to update just use the PUT instead which provides you only update what field you like.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] create complete address of an existing user.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Insert new address, postal code isn't required.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddressInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.RespWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AddressRespData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dev/users/{id}/changepw": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Attempt change password from an existing user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] change an existing user password",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Entry existing user valid credentials and the new one.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangePwInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.NormalResp"
                        }
                    }
                }
            }
        },
        "/dev/users/{id}/profile": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get complete details of logged user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] get complete profile of an existing user.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserCompleteDataResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Change an existing user profile information which consists of email, fullname, and username. Email format following RFC 5322 format. For update address info instead, please use PUT user/profile/address instead. This endpoint has not POST method which same behaviour already handled by POST /register",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] change an existing user profile information.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Only insert profile aspect need to be updated. Inserted value may lead to error for some reasons such updating to used username",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProfileInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.NormalResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "This request lead to removing all user's related data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] delete an existing user role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.RespWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UserDataResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dev/users/{id}/role": {
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Every patch request to this endpoint lead to switch role between [customer], [merchant]. Be careful, switching from [merchant] to [customer] lead to wipe out all user's merchant data (also its products)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev/User"
                ],
                "summary": "[EXPERIMENTAL] [RESTRICTED] change an existing user role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_jwt_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.RespWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RoleDataResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AddressInput": {
            "type": "object",
            "required": [
                "address_line",
                "city",
                "country",
                "phone_number"
            ],
            "properties": {
                "address_line": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "integer"
                },
                "postal_code": {
                    "type": "integer"
                }
            }
        },
        "models.AddressInputNotBinding": {
            "type": "object",
            "properties": {
                "address_line": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "integer"
                },
                "postal_code": {
                    "type": "integer"
                }
            }
        },
        "models.AddressRespData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "postal_code": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ChangePwInput": {
            "type": "object",
            "required": [
                "new_password",
                "password"
            ],
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.RoleDataResp": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UpdateProfileInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserCompleteDataResp": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/models.AddressRespData"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserDataResp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.NormalResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "string"
                }
            }
        },
        "utils.RespWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}